[platforms/drm] Properly adjust cursor position on a rotated output
Summary:
This change refactors the way how the cursor position on a screen is
calculated. It is changed to a matrix multiplication with the
adjustments for the rotated screen just being additional steps added to
the matrix.

With this calculation I got correct positions for all variants of
scaled, rotated and differently positioned outputs.

Reviewers: #kwin, #plasma

Subscribers: plasma-devel, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D8627