Do not call into NETWinInfo in AbstractClient::setSkipPager
Summary:
There's a call to a virtual doSetSkipPager for performing this call and
Client implements it with the code which now is removed.

This fixes a crash when using setSkipPager for a Wayland window.

Test Plan:
Crash which happens with a new test case (see dependent phab
request) is fixed.

Reviewers: #kwin, #plasma

Subscribers: plasma-devel, kwin

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9865